files cleaning
